+++ smapy/actions/__init__.py
@@ -1,4 +1,0 @@
-from smapy.utils import find_submodules
-
-modules = find_submodules(__name__)
-__all__ = modules.keys()