+++ tests/test_runnable.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import copy
 import datetime
 import json
@@ -21,7 +23,7 @@ class TestRunnableMeta(TestCase):
         class Test(Base):
             pass
 
-        self.assertEqual(Test.name, 'runnable_test.Test')
+        self.assertEqual(Test.name, 'test_runnable.Test')
 
     def test_custom_name(self):