+++ HISTORY.md
@@ -1,5 +1,12 @@
 # History
 
+## 0.0.2
+
+* Add `smapy` CLI
+* Add default actions ands resources
+* Import actions dynamically
+* Update project structure
+
 ## 0.0.1
 
 * First release on PyPI.